web-report-front-back/web-report/requirements.txt
django==2.2.5
django-tastypie==0.14.2